Add new generic Packstack and puppet-openstack-integration playbooks
There is no real purpose in maintaining a different playbook per
test scenario when the only thing that changes in them is the name
of the test that will be run.
As such, re-work the playbooks to be generic and validate that a
test scenario is properly passed as an extra var.

Change-Id: I2f3b7bf8782f395e0d4b74c360aa5cea72441f28