Update patch set 1
Patch Set 1:

> We seem to be having an issue with centos8-based jobs, because we
 > are still using centos 8.2-based images.
 > 
 > We need to get https://review.opendev.org/c/openstack/diskimage-builder/+/765963
 > merged in diskimage-builder so we can fix it.

Yeah, I was proposing to enable "powertools" between those two tasks [1] to unblock promotion, but it's kind of workaround. After the update task, the host has new repoid, but not before, that's why it does not know 'powertools' here.
So, you're right, it's better to fix it at the source.

[1] https://github.com/rdo-infra/ansible-role-weirdo-common/blob/b295724fe0d4f73dc48449b8b2a410b7115d09aa/tasks/setup.yml#L17#L29

Patch-set: 1