Add container-build scenario000
We set "/var/log/weirdo-project" as the workspace directory
for the build_containers role in order to retrieve the logs
with "logs" weirdo role.
We want to use the repositories set by weirdo common role, so
we explicitly set cifmw_build_containers_repo_dir to
'/etc/yum.repos.d' (default is "{{ cifmw_build_containers_basedir
}}/artifacts/repositories" [1]).

Also, we remove the roles to be inspected by ansible-lint, as
it's out of the scope of this project. Plus, those roles are
already inspected by the linter on their check jobs. I replaced
obsolete jenkins/ directory by the roles/ one.

[1] https://github.com/openstack-k8s-operators/ci-framework/blob/f3c77a1b8e24eeedd4654faedea3998802566091/ci_framework/roles/build_containers/defaults/main.yml#L48

Change-Id: Id4ff1b1d4eb6947bc080af0471f389458e8a324e